tests: Extract GetBundleFromConfigMap to libinfra
Move cert related functions to a dedicated file [0].

[0]
https://github.com/kubevirt/kubevirt/commit/3edd103d4adeb1d8fb01b3d4c600d9076bc5a1f6
