+++ tests/libinfra/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//pkg/certificates/triple/cert:go_default_library",
         "//pkg/downwardmetrics/vhostmd/api:go_default_library",
         "//pkg/virt-controller/leaderelectionconfig:go_default_library",
+        "//pkg/virt-operator/resource/generate/components:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//tests:go_default_library",
