+++ tests/libinfra/certificates.go
@@ -20,11 +20,18 @@
 package libinfra
 
 import (
+	"context"
+	"crypto/x509"
 	"reflect"
 
 	"github.com/onsi/gomega"
+	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"kubevirt.io/kubevirt/pkg/certificates/triple/cert"
+	"kubevirt.io/kubevirt/pkg/virt-operator/resource/generate/components"
+
+	"kubevirt.io/kubevirt/tests/flags"
+	"kubevirt.io/kubevirt/tests/framework/kubevirt"
 )
 
 func ContainsCrt(bundle []byte, containedCrt []byte) bool {
@@ -39,4 +46,16 @@ func ContainsCrt(bundle []byte, containedCrt []byte) bool {
 		}
 	}
 	return attached
+}
+
+func GetBundleFromConfigMap(configMapName string) ([]byte, []*x509.Certificate) {
+	virtClient := kubevirt.Client()
+	configMap, err := virtClient.CoreV1().ConfigMaps(flags.KubeVirtInstallNamespace).Get(context.Background(), configMapName, v1.GetOptions{})
+	gomega.Expect(err).ToNot(gomega.HaveOccurred())
+	if rawBundle, ok := configMap.Data[components.CABundleKey]; ok {
+		crts, err := cert.ParseCertsPEM([]byte(rawBundle))
+		gomega.Expect(err).ToNot(gomega.HaveOccurred())
+		return []byte(rawBundle), crts
+	}
+	return nil, nil
 }
